+++ CHANGELOG.md
@@ -15,6 +15,7 @@ For updates follow [@implydata](https://twitter.com/implydata) on Twitter.
 - Ability to [generate programmatic links](docs/generating-links-into-pivot.md) into Pivot
 - Ability to configure `serverHost`
 - Auto refresh icon rotation
+- Pivot is npm shrinkwrapped
 
 ## 0.8.42